Key Pulldown governs Notenames
Key pulldown governs what the notenames are displayed.  E.G. if you pull
down to key of G, you will see all the note names have sharps (G has one
sharp).  If you select the notegroup Bb Dorian: in INTERVAL mode, you
will see Bb labeled as the root.  In NOTES mode, you will see A# colored
as the root.  If you then use the key pulldown to F, you will see the
root always labeled as Bb.  This is because the key of F is in flats.